migrate to generateQualifiedIds
